Mobile App Item 127: Member Online Tee-Time bookings (‘MOTT’)
- Implements logical functionality of Buggy Feature [Done]
- UPDATE Alert message When make or cancel booking [Done]
- Deploy MHS GOLF club build with MOTT [Done]

+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/web/models/teetimebooking/getdaydata/Product.java
@@ -16,7 +16,7 @@ public class Product {
     private String Description;
     @SerializedName("Price")
     @Expose
-    private float Price;
+    private int Price;
     @SerializedName("BuggyIsValid")
     @Expose
     private Boolean BuggyIsValid;
@@ -52,11 +52,11 @@ public class Product {
         this.Description = Description;
     }
 
-    public float getPrice() {
+    public int getPrice() {
         return Price;
     }
 
-    public void setPrice(float Price) {
+    public void setPrice(int Price) {
         this.Price = Price;
     }
